+++ ItemQualities/Assets/ItemQualities/Scripts/Items/FragileDamageBonus.cs
@@ -1,4 +1,6 @@
-﻿using MonoMod.Cil;
+﻿using ItemQualities.Utilities.Extensions;
+using Mono.Cecil.Cil;
+using MonoMod.Cil;
 using R2API;
 using RoR2;
 
@@ -36,18 +38,61 @@ namespace ItemQualities.Items
         {
             ILCursor c = new ILCursor(il);
 
-            if (!c.TryFindNext(out ILCursor[] foundCursors,
-                               x => x.MatchLdfld<HealthComponent.ItemCounts>(nameof(HealthComponent.ItemCounts.fragileDamageBonus)),
-                               x => x.MatchCallOrCallvirt(typeof(HealthComponent).GetProperty(nameof(HealthComponent.isHealthLow)).GetMethod),
+            int watchItemTransformationVarIndex = -1;
+            int watchItemTransformationResultVarIndex = -1;
+            if (!c.TryGotoNext(MoveType.After,
+                               x => x.MatchLdfld<HealthComponent.ItemCounts>(nameof(HealthComponent.ItemCounts.fragileDamageBonus))) ||
+                !c.TryGotoNext(MoveType.After,
+                               x => x.MatchLdloca(typeof(Inventory.ItemTransformation), il, out watchItemTransformationVarIndex),
+                               x => x.MatchLdarg(0),
+                               x => x.MatchLdfld<HealthComponent>(nameof(HealthComponent.body)),
+                               x => x.MatchCallOrCallvirt<CharacterBody>("get_" + nameof(CharacterBody.inventory)),
+                               x => x.MatchLdloca(typeof(Inventory.ItemTransformation.TryTransformResult), il, out watchItemTransformationResultVarIndex),
                                x => x.MatchCallOrCallvirt<Inventory.ItemTransformation>(nameof(Inventory.ItemTransformation.TryTransform))))
             {
-                Log.Error("Failed to find watch break location");
+                Log.Error("Failed to find patch location");
                 return;
             }
 
-            c.Goto(foundCursors[2].Next, MoveType.After); // call Inventory.ItemTransformation.TryTransform
+            c.Emit(OpCodes.Ldarg_0);
+            c.Emit(OpCodes.Ldloca, watchItemTransformationVarIndex);
+            c.Emit(OpCodes.Ldloca, watchItemTransformationResultVarIndex);
+            c.EmitDelegate<ConsumeQualityWatchesDelegate>(consumeQualityWatches);
 
-            ItemHooks.EmitCombinedQualityItemTransformationPatch(c);
+            static bool consumeQualityWatches(bool result, HealthComponent healthComponent, in Inventory.ItemTransformation itemTransformation, ref Inventory.ItemTransformation.TryTransformResult consumeTransformResult)
+            {
+                CharacterBody body = healthComponent ? healthComponent.body : null;
+                Inventory inventory = body ? body.inventory : null;
+
+                if (inventory)
+                {
+                    for (QualityTier qualityTier = 0; qualityTier < QualityTier.Count; qualityTier++)
+                    {
+                        Inventory.ItemTransformation qualityItemTransformation = itemTransformation;
+                        qualityItemTransformation.originalItemIndex = ItemQualitiesContent.ItemQualityGroups.FragileDamageBonus.GetItemIndex(qualityTier);
+                        qualityItemTransformation.newItemIndex = ItemQualitiesContent.ItemQualityGroups.FragileDamageBonusConsumed.GetItemIndex(qualityTier);
+
+                        if (qualityItemTransformation.TryTransform(inventory, out Inventory.ItemTransformation.TryTransformResult qualityWatchConsumeTransformationResult))
+                        {
+                            result = true;
+
+                            static void addStackValues(ref Inventory.ItemStackValues a, in Inventory.ItemStackValues b)
+                            {
+                                a.permanentStacks += b.permanentStacks;
+                                a.temporaryStacksValue += b.temporaryStacksValue;
+                                a.totalStacks += b.totalStacks;
+                            }
+
+                            addStackValues(ref consumeTransformResult.takenItem.stackValues, qualityWatchConsumeTransformationResult.takenItem.stackValues);
+                            addStackValues(ref consumeTransformResult.givenItem.stackValues, qualityWatchConsumeTransformationResult.givenItem.stackValues);
+                        }
+                    }
+                }
+
+                return result;
+            }
         }
+
+        delegate bool ConsumeQualityWatchesDelegate(bool result, HealthComponent healthComponent, in Inventory.ItemTransformation itemTransformation, ref Inventory.ItemTransformation.TryTransformResult consumeTransformResult);
     }
 }