+++ ItemQualities/Assets/ItemQualities/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## ?.?.?
 
+* Fixed Power Elixir quality effect not working properly.
+
 * Fixed some qualities not being sorted in logbook.
 
 * Fixed incompatibility with RiskyTweaks if `Frost Relic - Remove Bubble` config was enabled.